Move result_type_signature_helper in a helper file
